Darnton experiment complete
The functions written for the illegal books experiment have been debugged and moved into a python module. The Darnton data was wrangled in R and deduped in Python. (The csvs have been kept out of the repo for licensing reasons.) Early results from both experiments are encouraging. The next step will be to conduct an error analysis on the output.